Added onReceive( ) and available( )
Tested with Teensy 3.2, 3.5 and 3.6.
The onReceive function can be called to set up a callback function which gets called by the interrupt handler to pass incoming data to the application. It should not be used in combination with the transfer( ) function (and friends). This is mostly useful when a SPI port is set up in slave mode.

When the callback function is called, it gets the user data parameter from the onReceive function as one of the parameters. This should make it easy for applications to keep track of their data. Each SPIClass instance keeps track of its own callback and userdata.

To cancel receive callbacks, use onReceive(NULL, NULL);

The available( ) function can be used to check if any incoming data is available. This is mostly useful in slave mode, and probably only when the application doesn't use onReceive.