Implemented AVR version of SPI slave mode
Tested with an Arduino Duemilenova clone.

I'm not 100% happy with this. The ISR is not fast enough to even do the callback without losing data. So I implemented a FIFO buffer that gets processed by the initial interrupt, while subsequent (recursive) interrupts are permitted to add data to the FIFO.

The stack is probably not going to overflow because of the nested interrupts (though that's difficult to prove), but it's impossible to predict how big the FIFO buffer needs to be. It depends on the application, mostly the maximum message size and the time between messages.

I also don't like how basically because of the way Arduino interrupts work, the code will always be in the application even if it's not required (e.g. a SPI master application). For a Teensy this is less critical because they have tons of memory but in an Atmega 328P, not so much.